Refactored Add functionality
