+++ Cargo.toml
@@ -24,7 +24,6 @@ serde_derive = "1.0.144"
 toml = "0.5.9"
 serde_json = "1.0"
 reqwest = { version = "0.11", features = ["blocking", "json"] } 
-typer = "0.1.1"
 
 [dev-dependencies]
 tempfile = "3.0.4"
