+++ src/huak/package/metadata.rs
@@ -1,15 +1,14 @@
-
-use serde_derive::{Serialize, Deserialize};  
+use serde_derive::{Deserialize, Serialize};
 
 // This models the response structure - check https://pypi.org/pypi/typer/json - typer is the package
 #[derive(Serialize, Deserialize, Debug)]
-pub struct PyPi{
+pub struct PyPi {
     pub info: PyPiInfo,
     last_serial: u64,
     releases: serde_json::value::Value,
     urls: Vec<serde_json::value::Value>,
     vulnerabilities: Vec<serde_json::value::Value>,
-} 
+}
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PyPiInfo {
@@ -33,11 +32,10 @@ pub struct PyPiInfo {
     pub project_url: String,
     pub project_urls: serde_json::value::Value,
     pub release_url: String,
-    pub requires_dist: serde_json::value::Value, 
-    pub requires_python: String, 
-    pub summary: String, 
+    pub requires_dist: serde_json::value::Value,
+    pub requires_python: String,
+    pub summary: String,
     pub version: String,
     pub yanked: bool,
-    pub yanked_reason: serde_json::value::Value
+    pub yanked_reason: serde_json::value::Value,
 }
-
