+++ src/huak/package/mod.rs
@@ -1,5 +1,5 @@
 mod build; // TODO
+pub mod metadata;
 mod publish; // TODO
 /// Python packaging namespace.
 pub mod python;
-pub mod metadata;
